Created a class called SortingAlgorithms
in this class, you have a method called FindSmallest which is used to find the smallest number in an array that has been passed as an argument. it also finds the index of that smallest number. so normally with the way the python code was written in the Grokking Algorithms book, all i just needed to be returned was the smallest index and then when you move over to the second method in the class which is called SelectionSort(more on that in the next paragraph), you remove the element at that particular index, but you also run into problems when you try to now use that same index in adding an element. In using ArrayLists, to remove, all you need is the index while to add an element, you need the element and not the index. (although now that i think about it, it could be simplified. I should commit this bit first)

in this class, you also have a method which was mentioned in ditto called SelectionSort and there was a lot of issues with this one too which i ended up resolving. i had to introduce a variable in ln21 called unsortedArraySize which held the constant amount of the unsortedArray. without that, the returned sortedArray just kept on getting smaller because the condition in the for loop gets the size of the unsortedArray but it has already been reduced and that was a problem because you get half of the Array.
I left the while loop in the comment in this method because it also works and i like it.