Made changes to the methods of the SortingAlgorithms class
when i was writing the message of the last commit, an idea came to me or came to me again because i think it might have been the idea that i had when i was doing laundry. And this has led me to change how the methods work. So in the case of findSmallest, instead of returning an array which contains the smallest number and its index, all i just need to do is return the index and then use the get method for ArrayList in the selectionSort method to fetch the number at that particular index in the unsorted Array and then add it to the sortedArray and then removing it from the unsortedArray instead of removing the number first from the unsortedArray and then adding it to the sortedArray. the latter creates problems as evidenced by the tests.