created the second selectionSort method(descending)
i created a method that does selection sort but instead of in ascending order, it is in descending order. this wasn't much of a stress. i didn't really have to do anything new. i just had to change signs and variable names. i made use of a while loop here rather than a for loop like i did in selectionSort(ascending).  i also wrote a test to test the method too.