Worked on ex 4.1 in the Grokking Algorithms book
using recursion to sum up the numbers in an array was explained and the task of ex4.1 was writing out the code. i just did. i wrote tests to test it as well to see if it worked, and it sure does. Recursion is a really interesting idea.

i also added a comment denoting my solution to ex4.3. i had written the solution earlier but had no idea that it would be an exercise later on.